Fix braindead mistake in flip
The flipPoint function was mis-transcribed, causing piece flipping to
result in different points, but the same effective shape.

Added a solver unit test that will only pass with a flip. Also, adjusted
the verbose mode to show all pieces (with flips and rotations) since
that inspection is what originally made the bug obvious.